requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
pylint>=3.0.0
python-dotenv>=1.0.0
langchain-groq>=0.1.0
